requirement.txt
# Core dependencies for ImgiNav
torch>=2.1.0
torchvision>=0.16.0
torchaudio>=2.1.0

# Data processing
numpy>=1.24.0
pandas>=2.0.0
pillow>=10.0.0
pyarrow>=12.0.0

# Visualization
matplotlib>=3.7.0
opencv-python>=4.8.0

# Scientific computing
scipy>=1.11.0
scikit-learn>=1.3.0
scikit-image>=0.21.0

# Configuration and utilities
pyyaml>=6.0
tqdm>=4.65.0
jsonschema>=4.17.0

# 3D processing
trimesh>=3.23.0
pyglet>=2.0.0

# Development tools
ipython>=8.14.0
jupyter>=1.0.0
jupyterlab>=4.0.0

# Additional pip dependencies
tensorboard
wandb
python-dotenv
rich
typer
